requirements.txt
pytest-testinfra
pyyaml
